separate css and components from ts files
