+++ js/website/src/renderLogo.tsx
@@ -1,4 +1,4 @@
 import { renderToString } from "react-dom/server";
-import { Logo } from "./Logo.tsx";
+import { Logo } from "./components/Logo";
 
 console.log(renderToString(<Logo colors={["#3e4756"]} />));
